+++ src/tUilKit/output/zorder/__init__.py
@@ -1,0 +1,4 @@
+"""zorder sub-package: z-index ordering for layered windows."""
+from tUilKit.output.zorder.zorder import ZOrderManager
+
+__all__ = ["ZOrderManager"]